Moved pull implementation from stream.hpp to new file pull.hpp
This refactor required removal of pull.toStream(), which only invoked stream's constructor.

pull.toStream() was only used in a few places inside stream, and I felt that it added an unnecessary cyclic dependency between the two classes.
